refactor: 패키지명 domain->entity
