+++ src/main/java/datastreams_knu/bigpicture/news/entity/News.java
@@ -1,4 +1,4 @@
-package datastreams_knu.bigpicture.news.domain;
+package datastreams_knu.bigpicture.news.entity;
 
 import datastreams_knu.bigpicture.common.domain.BaseEntity;
 import jakarta.persistence.*;
